dashboard/requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==4.2.10
django-cors-headers==4.3.1
django-crispy-forms==2.1
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
